+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Move template : Add journal field to wizard
 - Move template : Add new field description
 - HR : Added a leave line configuration menu in leave management
+- Move Template : Add totals for amount type
 - Move template : Add details button to grid view to display fields
 - Move template : Add change track on update
 - EMPLOYEE : set seniorityDate by hireDate